requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
prometheus-fastapi-instrumentator